reorganisation structure input
