+++ packages/react/components/input/gauge/InputGauge.tsx
@@ -4,7 +4,7 @@ import React from 'react'
 import { hashClass } from '../../../helpers'
 import { Icon, IconName, IconSize } from '../../icon'
 import { IValidationRules } from '../InputProps'
-import { useGauge } from '../hook/useGauge'
+import { useGauge } from './hook/useGauge'
 
 interface InputGaugeProps {
   validationRules?: IValidationRules
